General edits
Began on session handling - once user signs up or logs in, query param is brought to react app and is parsed to store current user uuid in sessionStorage and as a session cookie. LOGIN button still doesn't do shit - gonna work on that. Keeping user uuid on frontend is prob not secure but will make querying backend much easier