Big changes
reworked some style changes to the ticket portal - added some more checks and referral links to make sure user is signed in. implemented the start of what will be a guest view, where the user may want to see an example of the application without signing in or making account. WIP